✨ Create geojson de-duplicated db in insertion step

+++ Database/insert_events.py
@@ -2,6 +2,7 @@ import argparse
 import os
 import pathlib
 import sqlite3
+from typing import Any
 
 import pandas as pd
 from pandarallel import pandarallel
@@ -17,22 +18,10 @@ from Database.scr.normalize_utils import (
 pandarallel.initialize(progress_bar=False, nb_workers=5)
 
 if __name__ == "__main__":
-    event_levels = {
-        "l1": {
-            "prefix": "Total_Summary",
-            "location_columns": {"Administrative_Areas": list},
-        },
-        "l2": {
-            "prefix": "Instance_Per_Administrative_Areas_",
-            "location_columns": {"Administrative_Areas": list},
-        },
-        "l3": {
-            "prefix": "Specific_Instance_Per_Administrative_Area_",
-            "location_columns": {
-                "Administrative_Area": str,
-                "Locations": list,
-            },
-        },
+    event_levels: dict[str, dict[str, str]] = {
+        "l1": {"Administrative_Areas": "list"},
+        "l2": {"Administrative_Areas": "list"},
+        "l3": {"Administrative_Area": "str", "Locations": "list"},
     }
     parser = argparse.ArgumentParser()
     parser.add_argument(
@@ -112,25 +101,56 @@ if __name__ == "__main__":
         type=str,
     )
 
+    parser.add_argument(
+        "-s",
+        "--save_output",
+        dest="save_output",
+        action="store_true",
+        help=f"Pass to replace store output. Can be combined with `--dump_geojson_to_file` and `--dry_run`to generate a .parquet copies of the de-duplictaed geojson output without doing any database insertion.",
+        required=False,
+    )
+
+    parser.add_argument(
+        "-o",
+        "--output_dir",
+        dest="output_dir",
+        help="Pass an output directory for storing the data as identical .parquet files with de-duplicated geojson output.",
+        required=False,
+    )
+
+    parser.add_argument(
+        "-d",
+        "--dry_run",
+        dest="dry_run",
+        action="store_true",
+        help="Pass to start a dry run that does not insert anything into the database. Useful with `--save_output`",
+        required=False,
+    )
+
+    logger = Logging.get_logger(f"database-insertion", level="INFO")
     args = parser.parse_args()
-    logger = Logging.get_logger(f"database-insertion", level="INFO", filename="v1_full_run_insertion_raw.log")
+    logger.info(f"ARGS: {args}")
     utils = NormalizeUtils()
     validation = CategoricalValidation()
 
-    connection = sqlite3.connect(args.database_name)
-    cursor = connection.cursor()
-    files = os.listdir(args.file_dir)
+    if not args.dry_run:
+        connection = sqlite3.connect(args.database_name)
+        cursor = connection.cursor()
+
     errors = pd.DataFrame()
+    files = os.listdir(args.file_dir)
+    if args.save_output:
+        assert args.output_dir, "Must specify an output directory when `--save_output` is selected"
+        pathlib.Path(args.output_dir).mkdir(parents=True, exist_ok=True)
 
     # levels
     main_level = "l1"
     sub_levels = [x for x in event_levels.keys() if x != "l1"]
+    geojson_utils = GeoJsonUtils(nid_path=args.nid_path)
 
-    if args.dump_geojson_to_file:
-        geojson_utils = GeoJsonUtils(nid_path=args.nid_path)
     if args.event_level == main_level:
         args.target_table = "Total_Summary"
-        logger.info(f"Inserting {main_level}...\n")
+        logger.info(f"Processing {main_level}...\n")
         for f in tqdm(files, desc="Files"):
             data = pd.read_parquet(f"{args.file_dir}/{f}", engine="fastparquet")
             data = utils.replace_nulls(data)
@@ -140,7 +160,7 @@ if __name__ == "__main__":
 
             # geojson in l1 is always of type list
             if args.dump_geojson_to_file:
-                for col in event_levels[args.event_level]["location_columns"].keys():
+                for col in event_levels[args.event_level].keys():
                     logger.info(f"Processing GeoJson column {col}_GeoJson in {args.event_level}; File: {f}")
 
                     data[f"{col}_GeoJson"] = data.parallel_apply(
@@ -156,40 +176,41 @@ if __name__ == "__main__":
                         axis=1,
                     )
 
-            logger.info("Converting everything to strings...")
-            data.replace(float("nan"), None, inplace=True)
-            for c in data.columns:
-                data[c] = data[c].astype(str)
+            if args.save_output:
+                data.to_parquet(f"{args.output_dir}/{f}", engine="fastparquet")
 
             # change if_exists to "append" to avoid overwriting the database
             # choose "replace" to overwrite the database with a fresh copy of the data
-            for i in tqdm(range(len(data)), desc=f"Inserting {args.event_level} into {args.database_name}"):
-                try:
-                    data.iloc[i : i + 1].to_sql(
-                        name=args.target_table,
-                        con=connection,
-                        if_exists=args.method,
-                        index=False,
-                    )
-                except sqlite3.IntegrityError as err:
-                    logger.debug(
-                        f"""Could not insert event for level {args.event_level}. Error {err}.
-                                The problematic row will be stored in /tmp/ with the error. GeoJson columns will not be included."""
-                    )
-                    err_row = data.iloc[i : i + 1][[x for x in data.columns if "GeoJson" not in x]].copy()
-                    err_row["ERROR"] = err
-                    errors = pd.concat([errors, err_row], ignore_index=True)
+            if not args.dry_run:
+                for i in tqdm(range(len(data)), desc=f"Inserting {f} into {args.database_name}"):
+                    try:
+                        data.iloc[i : i + 1].to_sql(
+                            name=args.target_table,
+                            con=connection,
+                            if_exists=args.method,
+                            index=False,
+                        )
+                    except sqlite3.IntegrityError as err:
+                        logger.debug(
+                            f"""Could not insert event for level {args.event_level}. Error {err}.
+                                    The problematic row will be stored in /tmp/ with the error. GeoJson columns will not be included."""
+                        )
+                        err_row = data.iloc[i : i + 1][[x for x in data.columns if "GeoJson" not in x]].copy()
+                        err_row["ERROR"] = err
+                        errors = pd.concat([errors, err_row], ignore_index=True)
 
     elif args.event_level in sub_levels:
-        logger.info(f"Inserting {args.event_level}...\n")
-        assert args.target_table, f"When inserting sublevels ({sub_levels}), the target table must be specified!"
+        logger.info(f"Processing level {args.event_level}: {args.target_table}...\n")
+        if not args.dry_run:
+            logger.info(f"Inserting {args.event_level}...\n")
+            assert args.target_table, f"When inserting sublevels ({sub_levels}), the target table must be specified!"
 
-        check_table = cursor.execute(
-            f"""SELECT name FROM sqlite_master WHERE type='table'
-                AND name='{args.target_table}'; """
-        ).fetchall()
+            check_table = cursor.execute(
+                f"""SELECT name FROM sqlite_master WHERE type='table'
+                    AND name='{args.target_table}'; """
+            ).fetchall()
 
-        assert len(check_table) == 1, f"Table name {args.target_table} incorrect! Found {check_table} instead."
+            assert len(check_table) == 1, f"Table name {args.target_table} incorrect! Found {check_table} instead."
 
         for f in tqdm(files, desc="Files"):
             data = pd.read_parquet(f"{args.file_dir}/{f}", engine="fastparquet")
@@ -197,9 +218,10 @@ if __name__ == "__main__":
 
             logger.info(f"Popping GeoJson files out for level {args.event_level} and onto disk")
             if args.dump_geojson_to_file:
-                for col, _type in event_levels[args.event_level]["location_columns"].items():
+                for col, _type in event_levels[args.event_level].items():
+                    eval_type: Any = eval(_type)
                     logger.info(f"Processing GeoJson column {col} in {args.event_level}; File: {f}")
-                    if _type == list:
+                    if eval_type == list:
                         data[f"{col}_GeoJson"] = data.parallel_apply(
                             lambda row: (
                                 [
@@ -213,50 +235,51 @@ if __name__ == "__main__":
                             axis=1,
                         )
 
-                    elif _type == str:
+                    elif eval_type == str:
                         data[f"{col}_GeoJson"] = data.parallel_apply(
                             lambda row: (
                                 geojson_utils.geojson_to_file(row[f"{col}_GeoJson"], row[f"{col}_Norm"])
-                                if isinstance(row[f"{col}_GeoJson"], str)
+                                if row[f"{col}_GeoJson"]
                                 else None
                             ),
                             axis=1,
                         )
 
+            if args.save_output:
+                data.to_parquet(f"{args.output_dir}/{f}", engine="fastparquet")
+
             # change if_exists to "append" to avoid overwriting the database
             # choose "replace" to overwrite the database with a fresh copy of the data
-            logger.info("Converting everything to strings...")
-            data.replace(float("nan"), None, inplace=True)
-            for c in data.columns:
-                data[c] = data[c].astype(str)
-
-            for i in tqdm(range(len(data)), desc=f"Inserting {args.event_level} into {args.database_name}"):
-                try:
-                    data.iloc[i : i + 1].to_sql(
-                        name=args.target_table,
-                        con=connection,
-                        if_exists=args.method,
-                        index=False,
-                    )
-                except sqlite3.IntegrityError as err:
-                    logger.debug(
-                        f"""Could not insert event for level {args.event_level}. Error {err}.
-                                The problematic row will be stored in tmp/ with the error. GeoJson columns will not be included."""
-                    )
-                    err_row = data.iloc[i : i + 1][[x for x in data.columns if "GeoJson" not in x]].copy()
-                    err_row["ERROR"] = err
-                    errors = pd.concat([errors, err_row], ignore_index=True)
+            if not args.dry_run:
+                for i in tqdm(range(len(data)), desc=f"Inserting {f} into {args.database_name}"):
+                    try:
+                        data.iloc[i : i + 1].to_sql(
+                            name=args.target_table,
+                            con=connection,
+                            if_exists=args.method,
+                            index=False,
+                        )
+                    except sqlite3.IntegrityError as err:
+                        logger.debug(
+                            f"""Could not insert event for level {args.event_level}. Error {err}.
+                                    The problematic row will be stored in tmp/ with the error. GeoJson columns will not be included."""
+                        )
+                        err_row = data.iloc[i : i + 1][[x for x in data.columns if "GeoJson" not in x]].copy()
+                        err_row["ERROR"] = err
+                        errors = pd.concat([errors, err_row], ignore_index=True)
 
     if args.dump_geojson_to_file:
         geojson_utils.store_non_english_nids()
 
-    if errors.shape != (0, 0):
-        from time import time
+    if not args.dry_run:
+        if errors.shape != (0, 0):
+            from time import time
 
-        tmp_errors_filename = (
-            f"{args.error_path}/db_insert_errors_{args.event_level}_{args.target_table}_{int(time())}.json"
-        )
-        logger.error(f"Insert errors were found! THIS ROW WAS NOT INSERTED! Storing in {tmp_errors_filename}")
-        pathlib.Path(args.error_path).mkdir(parents=True, exist_ok=True)
-        errors.to_json(tmp_errors_filename, orient="records", indent=3)
-    connection.close()
+            tmp_errors_filename = (
+                f"{args.error_path}/db_insert_errors_{args.event_level}_{args.target_table}_{int(time())}.json"
+            )
+            logger.error(f"Insert errors were found! THIS ROW WAS NOT INSERTED! Storing in {tmp_errors_filename}")
+            pathlib.Path(args.error_path).mkdir(parents=True, exist_ok=True)
+            errors.to_json(tmp_errors_filename, orient="records", indent=3)
+
+        connection.close()
